流程部署模块新增部署资源功能，通过配置流程文件获取流程资源信息并解释

+++ bc-workflow-core/src/main/java/cn/bc/workflow/deploy/service/DeployServiceImpl.java
@@ -142,7 +142,7 @@ public class DeployServiceImpl extends DefaultCrudService<Deploy> implements
 	public void dodeployCancel(Long deploymentId, Boolean isCascade) {
 		// 判断发布类型
 		Deploy deploy = this.deployDao.load(deploymentId);
-		if (null != deploy) {
+		if (null != deploy && null != deploy.getDeploymentId()) {
 			if (isCascade) {
 				repositoryService.deleteDeployment(deploy.getDeploymentId(),
 						true);// 级联删除流程
